+++ package_files/gogios.toml
@@ -74,10 +74,10 @@
 #   database = "gogios"
 
 
-# Output check data to a Sqlite3 database file
-[[databases.sqlite]]
-  ## Sqlite3 db path
-  db_file = "/var/lib/gogios/gogios.db"
+# # Output check data to a Sqlite3 database file
+# [[databases.sqlite]]
+#   ## Sqlite3 db path
+#   db_file = "/var/lib/gogios/gogios.db"
 
 
 ###########################
@@ -119,4 +119,3 @@
 #
 #   ## HTTP response timeout (default: 10s)
 #   response_timeout = "10s"
-
